Customer service

+++ fundtransfer/fundtransfer/src/main/java/com/example/fundtransfer/entity/Transaction.java
@@ -55,7 +55,7 @@ public abstract class Transaction {
 
     public void setToAccount(String toAccount) { this.toAccount = toAccount; }
 
-    public Double getAmount() { return amount; }
+    public BigDecimal getAmount() { return amount; }
 
     public void setAmount(BigDecimal amount) { this.amount = amount; }
